+++ wvlet-cli/src/main/scala/wvlet/lang/cli/WvletGlobalOption.scala
@@ -32,5 +32,5 @@ case class WvletGlobalOption(
       case _ =>
         error(s"Invalid log level pattern: ${p}")
   }
-
+  
 end WvletGlobalOption
