+++ wvlet-cli/src/main/scala/wvlet/lang/cli/WvletMain.scala
@@ -2,10 +2,13 @@ package wvlet.lang.cli
 
 import wvlet.airframe.launcher.{Launcher, argument, command, option}
 import wvlet.lang.BuildInfo
-import wvlet.lang.api.StatusCode
+import wvlet.lang.api.v1.query.QuerySelection.All
+import wvlet.lang.api.{StatusCode, WvletLangException}
 import wvlet.lang.catalog.Profile
-import wvlet.lang.compiler.codegen.GenSQL
-import wvlet.lang.compiler.{CompilationUnit, Compiler, CompilerOptions, WorkEnv}
+import wvlet.lang.cli.WvletMain.isInSbt
+import wvlet.lang.compiler.planner.ExecutionPlanner
+import wvlet.lang.compiler.*
+import wvlet.lang.runner.QuerySelector
 import wvlet.lang.server.{WvletServer, WvletServerConfig}
 import wvlet.log.LogSupport
 
@@ -13,6 +16,8 @@ object WvletMain:
   private def launcher: Launcher      = Launcher.of[WvletMain]
   def main(args: Array[String]): Unit = launcher.execute(args)
   def main(argLine: String): Unit     = launcher.execute(argLine)
+
+  def isInSbt: Boolean = sys.props.getOrElse("wvlet.sbt.testing", "false").toBoolean
 
 /**
   * 'wvlet' command line interface
@@ -29,50 +34,12 @@ class WvletMain(opts: WvletGlobalOption) extends LogSupport:
 
   @command(description = "Compile .wv files")
   def compile(compilerOption: WvletCompilerOption): Unit =
-    val currentProfile: Profile = Profile
-      .getProfile(compilerOption.profile, compilerOption.catalog, compilerOption.schema)
-
-    val compiler = Compiler(
-      CompilerOptions(
-        phases = Compiler.allPhases,
-        workEnv = WorkEnv(compilerOption.workFolder, opts.logLevel),
-        catalog = currentProfile.catalog,
-        schema = currentProfile.schema
-      )
-    )
-
-    val inputUnit =
-      (compilerOption.file, compilerOption.query) match
-        case (Some(f), None) =>
-          CompilationUnit.fromFile(s"${compilerOption.workFolder}/${f}".stripPrefix("./"))
-        case (None, Some(q)) =>
-          CompilationUnit.fromString(q)
-        case _ =>
-          throw StatusCode
-            .INVALID_ARGUMENT
-            .newException("Specify either --file or a query argument")
-
-    val compileResult = compiler.compileSingleUnit(inputUnit)
-
-    compileResult.reportAllErrors
-
-
-
-  end compile
+    try
+      WvletCompiler(opts, compilerOption).toSQL
+    catch
+      case e: WvletLangException =>
+        error(e.getMessage)
+        if !isInSbt then
+          System.exit(1)
 
 end WvletMain
-
-case class WvletCompilerOption(
-    @option(prefix = "-w", description = "Working folder")
-    workFolder: String = ".",
-    @option(prefix = "--profile", description = "Profile to use")
-    profile: Option[String] = None,
-    @option(prefix = "--catalog", description = "Context database catalog to use")
-    catalog: Option[String] = None,
-    @option(prefix = "--schema", description = "Context database schema to use")
-    schema: Option[String] = None,
-    @option(prefix = "--file", description = "Read a query from a file")
-    file: Option[String] = None,
-    @argument(description = "query")
-    query: Option[String] = None
-)
