+++ wvlet-cli/src/main/scala/wvlet/lang/cli/WvletREPL.scala
@@ -41,12 +41,11 @@ import scala.io.AnsiColor
 class WvletREPL(workEnv: WorkEnv, runner: WvletScriptRunner) extends AutoCloseable with LogSupport:
   import WvletREPL.*
 
-  private val useDumbTerminal = sys.props.getOrElse("wvlet.sbt.testing", "true").toBoolean
   private val terminal = TerminalBuilder
     .builder()
     .name("wvlet-shell")
     // Use dumb terminal for sbt testing
-    .dumb(useDumbTerminal).build()
+    .dumb(WvletMain.isInSbt).build()
 
   private val historyFile = new File(workEnv.cacheFolder, ".wv_history")
